Seperated the file into two files followed solid principles
